job/requirements.txt
sentence-transformers
boto3
pinecone-client
pandas
torch